Add acceptance tests.
Because the purpose of the mocha-eslint module is to run mocha tests which
check for ESLint errors, the acceptance tests must themselves also run such tests.

Therefore, the testing strategy is to execute a test file similar to what a consuming
module might use.  Basically, this means running a mocha test (mocha-eslint) inside
of mocha tests (the acceptance tests).

To do this, the stdout of the mocha-eslint tests need to be captured and then
checked for expected results.  The way this is accomplished is using the snippet
provided by @danielstjules in https://github.com/mochajs/mocha/issues/138#issuecomment-82727410

So far, the following tests are being performed:

- Verify that a test of a file which should not give ESLint warnings passes
- Verify that a test of a file which should give ESLint warnings does in fact fail
- Verify that a test of multiple paths works correctly.